Finally some love for the OSX port.  The Xcode project now compiles and runs, and contains only one Objective-C warning (more research required).  There are also some fixes for minor warnings.
There are still some issues, though.  First, SDL2.0.3 has to compiled
manually, as the public binaries don't work (this is documented and
a bug reported on the SDL mailing list).  Second, fullscreen->window
transition is causing a crash, which I still need to track down.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@2926 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba

+++ src/debugger/CartDebug.cxx
@@ -392,7 +392,7 @@ string CartDebug::disassemble(uInt16 start, uInt16 lines) const
     else
       buffer << "       ";
 
-    buffer << tag.disasm << setw(length - tag.disasm.length() + 2)
+    buffer << tag.disasm << setw(int(length - tag.disasm.length() + 2))
            << setfill(' ') << " "
            << setw(4) << left << tag.ccount << "   " << tag.bytes << endl;
   }
@@ -768,7 +768,7 @@ string CartDebug::loadListFile()
       if(addr_s.length() == 0)
         continue;
       const char* p = addr_s[0] == 'U' ? addr_s.c_str() + 1 : addr_s.c_str();
-      addr = strtoul(p, NULL, 16);
+      addr = (int)strtoul(p, NULL, 16);
 
       // For now, completely ignore ROM addresses
       if(!(addr & 0x1000))
